refactor: move _format_query_result tests to own file
Avoids exempting test_remote_management_cli.py at 519 lines — split
the standalone utility tests into test_format_query_result.py (47 lines)
so both files stay under 500.
